✨ Add realistic initial sync/syncRecords dance
JIRA: https://issues.jboss.org/browse/RAINCATCH-635

Currently a client will do a sync with the old hash, then if there's
changes on the server (different hash), it will do a syncRecords with
the clientRecs that it has (none initially for us). The server will
reply with a bunch of changes that the client needs to make at this
point. The client will then do another sync (with the same old hash
that it had) and then a syncRecords with a clientRecs object that
reflects the changes that the server told it that it should apply.

This change updates the portal and mobile flows to reflect that
to-and-fro.