🚧 Make a working mobile flow
JIRA: https://issues.jboss.org/browse/RAINCATCH-635

The changes here (of which there are many), give a working mobile
flow. By 'working', I mean that the steps don't collide with each
other, and the one result record is updated continuously. After
running this flow, the workorder should be marked as completed if you
look at it in the portal.